Extract ConnectionManager to a separate file
